SECURITY: be consistent about the hash algorithm
SECURITY doc was mentioning both SHA1 and SHA256. Although this is configurable the documentation should be consistent on which algorithms are used by default.
By looking at the `encrypt_and_sign` mentioned in the doc it uses `instance->crypto_hash_type`, which is set in `totem_get_crypto`, defaulting to `sha1` if not overriden.